refactor(address formatter): rename component
